Move features into own folder
