+++ lib/main.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
+import 'package:pizzaiolo_companion/src/features/settings/settings_controller.dart';
+import 'package:pizzaiolo_companion/src/features/settings/settings_service.dart';
 
 import 'src/app.dart';
-import 'src/settings/settings_controller.dart';
-import 'src/settings/settings_service.dart';
 import 'src/services/repository.dart';
 
 void main() async {
