+++ lib/src/app.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
+import 'package:pizzaiolo_companion/src/features/recipes/sample_item_details_view.dart';
+import 'package:pizzaiolo_companion/src/features/recipes/sample_item_list_view.dart';
+import 'package:pizzaiolo_companion/src/features/settings/settings_controller.dart';
+import 'package:pizzaiolo_companion/src/features/settings/settings_view.dart';
 
-import 'sample_feature/sample_item_details_view.dart';
-import 'sample_feature/sample_item_list_view.dart';
-import 'settings/settings_controller.dart';
-import 'settings/settings_view.dart';
 import 'services/repository.dart';
 
 /// The Widget that configures your application.
