+++ lib/src/features/recipes/sample_item_details_view.dart
