+++ lib/src/features/settings/settings_controller.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
+import 'package:pizzaiolo_companion/auth/secrets.dart';
 
 import 'settings_service.dart';
-
-import '../../auth/secrets.dart';
 
 /// A class that many Widgets can interact with to read user settings, update
 /// user settings, or listen to user settings changes.
